requirements.txt
Flask
GitHub-Flask
python-dotenv
langchain_core
langchain_community
langchain_classic
langchain-chroma
langchain_openai
duckduckgo-search
ddgs
gunicorn
gevent
supabase